Find Cousins in Family Tree
Class Member is a class that represents a single person in the family, and Class Family represents the whole family tree.
Suppose two people, P1 and P2, are i'th cousins. Let C1 be a child of P1 and C2 be a child of P2. Then, C1 is an i'th cousin of P2, 1 removed, and C2 is an i'th cousin of P1, 1 removed.

Let G1 now be a child of C1. G1 is an i'th cousin of P2, 2 removed.

In general, the type of cousin (what 'i' is) is the shorter distance to the ancestor of two people, and the amount removed is the difference between the distance to the common ancestor.